added table, intergrated navbar, added profile page
press "Profile" in top right to go to profile page
