+++ app/api/user/[id].test.tsx
@@ -5,6 +5,7 @@ import { GetUserTimeSheets } from './[id].tsx';
 
 
 const expected_user_json = "{\"body\":{\"task_sheet_ids\":[{\"id\":1},{\"id\":2}]}}";
+//const expected_user_json =  "{\"user\":{\"id\":2,\"task_sheet_ids\":\"[{ \\\"id\\\": 1 , tasks: \\\"[{task_id:1,task_name:\\\"homework1\\\",completed:false,user_ids:\\\"[{\\\\\\\"id\\\\\\\":1},\\\"+\\\"{\\\\\\\"id\\\\\\\":2}]\\\"},{task_id:2,task_name:\\\"homework2\\\",completed:true,user_ids:\\\"[{\\\\\\\"id\\\\\\\":1},\\\"+\\\"{\\\\\\\"id\\\\\\\":2}]\\\"}]\\\"},{ \\\"id\\\":2 , tasks: \\\"[{task_id:1,task_name:\\\"homework1\\\",completed:false,user_ids:\\\"[{\\\\\\\"id\\\\\\\":1},\\\"+\\\"{\\\\\\\"id\\\\\\\":2}]\\\"},{task_id:2,task_name:\\\"homework2\\\",completed:true,user_ids:\\\"[{\\\\\\\"id\\\\\\\":1},\\\"+\\\"{\\\\\\\"id\\\\\\\":2}]\\\"}]\\\"} ]\",\"first_name\":\"Patrick\",\"second_name\":\"Star\",\"student\":true}}";
 
 it('should return data with status 200', async () => {
